Added UnityRasa Breast Diagnosis Project: Verified Att UNet Seg Report
Leveraging my previous Unity Rasa UI project, but applying it for breast cancer diagnosis.

Also fixed an issue in this iteration of the project that was lurking in the previous
version, which was that now I can gpu scale the image that is displayed in the chat
window. This support is for both url and local images. In our case right now,
I tested it on the breast cancer segmentation qualitative report. Chatbot Rasa deploys
Attention UNet on backend after we make the request in Unity, then seg result report
is displayed in Unity